Web Inspector: Remove unused parameter of ScriptArguments::getFirstArgumentAsString https://bugs.webkit.org/show_bug.cgi?id=157301
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2016-05-03
Reviewed by Timothy Hatcher.

* inspector/ScriptArguments.cpp:
(Inspector::ScriptArguments::getFirstArgumentAsString):
* inspector/ScriptArguments.h:
Remove unused argument and related code.

* runtime/ConsoleClient.cpp:
(JSC::ConsoleClient::printConsoleMessageWithArguments):
Drive by remove unnecessary cast.

Canonical link: https://commits.webkit.org/175431@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@200398 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/JavaScriptCore/runtime/ConsoleClient.cpp
@@ -178,7 +178,7 @@ void ConsoleClient::printConsoleMessageWithArguments(MessageSource source, Messa
                 functionName = ASCIILiteral("(unknown)");
 
             StringBuilder callFrameBuilder;
-            callFrameBuilder.appendNumber(static_cast<unsigned long>(i));
+            callFrameBuilder.appendNumber(i);
             callFrameBuilder.appendLiteral(": ");
             callFrameBuilder.append(functionName);
             callFrameBuilder.append('(');
